Crate webb_relayer_config

Relayer Configuration Module 🕸️

A module for configuring the relayer.

Overview

The relayer configuration module is responsible for configuring the relayer. Possible configuration include:

port: The port the relayer will listen on. Defaults to 9955

evm: EVM based networks and the configuration. See config/config-6sided-eth-bridge for an example.

substrate: Substrate based networks and the configuration. See config/local-substrate for an example.

Modules

anchor

Generic anchor configuration

block_poller

Block poller configuration

cli

CLI configuration

defaults

Module for all the default values.

event_watcher

Event watcher configuration

evm

EVM configuration

signing_backend

Signing backend configuration

substrate

Substrate configuration

utils

Utils for processing configuration

Structs

EtherscanApiConfig

Configuration to add etherscan API key

FeaturesConfig

FeaturesConfig is the configuration for running relayer with option.

TxQueueConfig

TxQueueConfig is the configuration for the TxQueue.

UnlistedAssetConfig

UnlistedAssetConfig is the configuration for the assets that are not listed on any exchange.

WebbRelayerConfig

WebbRelayerConfig is the configuration for the webb relayer.